// old_src/types/TransitionType.ts
export interface ITransition {
  updatePart: HTMLElement[],
  step: (styles: ICSSStyleDeclaration, duration: number, timingFunction: string, delay: number, callBack?: () => void) => Transition;
  end: (callBack: () => void) => Transition;
  clear: () => Transition;
}

export type TTransitionItem = {
  styles: ICSSStyleDeclaration,
  duration?: number,
  timingFunction?: string,
  delay?: number,
  callBack?: () => void
};